_changes with doc_ids filter works for admin, added new unit tests for admin user, POST with GET params override, and GET request with _doc_ids filter
